+++ Gulpfile.js
@@ -2,7 +2,9 @@ var gulp = require('gulp');
 var wiredep = require('wiredep').stream;
 gulp.task('wiredep', function() {
     gulp.src('./_layouts/base.md')
-        .pipe(wiredep())
+        .pipe(wiredep({
+            ignorePath: '../'
+        }))
         .pipe(gulp.dest('./_layouts'));
 });