+++ assets/lib/bootstrap-ui-datetime-picker/example/index.js
@@ -8,14 +8,17 @@ app.controller('MyController', ['$scope', function($scope) {
     in10Days.setDate(in10Days.getDate() + 10);
 
     this.dates = {
-        date1: new Date('01 Mar 2015 00:00:00.000'),
-        date2: new Date(),
+        date1: new Date('2015-03-01T00:00:00Z'),
+        date2: new Date('2015-03-01T12:30:00Z'),
         date3: new Date(),
         date4: new Date(),
         date5: in10Days,
         date6: new Date(),
         date7: new Date(),
-        date8: new Date()
+        date8: new Date(),
+        date9: null,
+        date10: new Date('2015-03-01T09:00:00Z'),
+        date11: new Date('2015-03-01T10:00:00Z')
     };
 
     this.open = {
@@ -26,10 +29,13 @@ app.controller('MyController', ['$scope', function($scope) {
         date5: false,
         date6: false,
         date7: false,
-        date8: false
+        date8: false,
+        date9: false,
+        date10: false,
+        date11: false
     };
 
-    // Disable weekend selection
+    // Disable today selection
     this.disabled = function(date, mode) {
         return (mode === 'day' && (new Date().toDateString() == date.toDateString()));
     };
@@ -50,14 +56,11 @@ app.controller('MyController', ['$scope', function($scope) {
     };
 
     this.openCalendar = function(e, date) {
-        e.preventDefault();
-        e.stopPropagation();
-
         that.open[date] = true;
     };
 
     // watch date4 and date5 to calculate difference
-    this.calculateWatch = $scope.$watch(function() {
+    var unwatch = $scope.$watch(function() {
         return that.dates;
     }, function() {
         if (that.dates.date4 && that.dates.date5) {
@@ -69,6 +72,6 @@ app.controller('MyController', ['$scope', function($scope) {
     }, true);
 
     $scope.$on('$destroy', function() {
-        that.calculateWatch();
+        unwatch();
     });
 }]);